binksmachine/README.md
This is where we will explore various processes' of (the blockchain) in general a distributed ledger requires a peer-to-peer (P2P) computer network and consensus algorithms so that the ledger is reliably replicated across distributed computer nodes (servers, clients, etc.). We will create a form of distributed ledger technology specifically for our clients and partners (commonly associated with the Bitcoin cryptocurrency), which can either be on a public or private network. Infrastructure for data management is a common barrier to implementing DLT, which is where we will attempt to provide clarity for the future of our development of Web3 applications and utility for the above-mentioned groups of participants.